Move the quick button prefs repo into its own package
